requirements.txt
bleach
charset-normalizer
coreapi
coreschema
Django==4.2
django-admin-notification
django-bootstrap4
django-ckeditor
django-cors-headers
django-cuser
django-jazzmin
django-js-asset
django-modeltranslation
django-rest-swagger
django-taggit
django-tinymce
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
MarkupSafe
openapi-codec
pillow
PyJWT
PyMySQL
python-dateutil
PyYAML
requests
simplejson
sqlparse
stripe
typing_extensions
tzdata
uritemplate
urllib3
webencodings
django-import-export
uWSGI
pymysql
cryptography